Unit test va Integration test yozildi

+++ cashback-service/src/main/java/uz/nt/cashbackservice/service/Impl/CashbackCardServiceImpl.java
@@ -123,7 +123,6 @@ public class CashbackCardServiceImpl implements CashbackCardService {
                 Double updatedAmount = card.getAmount() - amount;
                 card.setAmount(updatedAmount);
                 addCashbackHistory(card.getId(), card.getAmount() - amount, amount, card.getAmount(), new Date());
-
                 message = messageSource.getMessage("operation.success", new String[]{}, locale);
                 return ResponseDto.<CashbackCardDto>builder()
                         .code(200)
@@ -192,46 +191,66 @@ public class CashbackCardServiceImpl implements CashbackCardService {
 
 
     @Override
-    public ResponseDto<CashbackCardDto> addCashback() {
+    public ResponseDto<CashbackCardDto> addCashback(HttpServletRequest request) {
+        Locale locale = request.getLocale();
+        String message;
         CashbackCard card = null;
         try{
             if(SecurityContextHolder.getContext().getAuthentication().getPrincipal() instanceof UserDto userDto){
                 Integer userId = userDto.getId();
                 card = CashbackCard.builder().amount(5000D).userId(userId).barcode(getBarcode()).build();
-                card = cashbackCardRepository.save(card);
+                cashbackCardRepository.save(card);
+
+                message = messageSource.getMessage("add.success", new String[]{}, locale);
+
+                return ResponseDto.<CashbackCardDto>builder()
+                        .success(true).code(200)
+                        .message(message)
+                        .responseData(cashbackMapper.toDto(card))
+                        .build();
             }
         }catch (Exception e){
+            message = messageSource.getMessage("error", new String[]{}, locale);
             return ResponseDto.<CashbackCardDto>builder()
                     .success(false).code(-1)
-                    .message("error".concat(": ").concat(e.getMessage()))
+                    .message(message.concat(": ").concat(e.getMessage()))
                     .responseData(null)
                     .build();
         }
+        message = messageSource.getMessage("operation.failed", new String[]{}, locale);
         return ResponseDto.<CashbackCardDto>builder()
-                .success(true).code(200)
-                .message("add.success")
-                .responseData(cashbackMapper.toDto(card))
+                .success(false).code(-1)
+                .message(message)
+                .responseData(null)
                 .build();
     }
 
 
     @Override
     public ResponseDto<Boolean> deleteCashBackCardById(Integer cashbackCardId, HttpServletRequest request) {
-
         Locale locale = request.getLocale();
         String message;
 
         if(cashbackCardId != null) {
-
-            cashbackCardRepository.deleteById(cashbackCardId);
-            message = messageSource.getMessage("operation.success", new String[]{}, locale);
+            if(cashbackCardRepository.existsById(cashbackCardId)) {
+                cashbackCardRepository.deleteById(cashbackCardId);
+                message = messageSource.getMessage("operation.success", new String[]{}, locale);
+                return ResponseDto.<Boolean>builder()
+                        .code(200)
+                        .responseData(true)
+                        .success(true)
+                        .message(message)
+                        .build();
+            }
+            message = messageSource.getMessage("not.found", new String[]{}, locale);
             return ResponseDto.<Boolean>builder()
-                    .code(200)
-                    .responseData(true)
-                    .success(true)
+                    .code(-2)
+                    .responseData(false)
+                    .success(false)
                     .message(message)
                     .build();
         }
+
         message = messageSource.getMessage("error", new String[]{}, locale);
         return ResponseDto.<Boolean>builder()
                 .code(-1)
@@ -248,13 +267,21 @@ public class CashbackCardServiceImpl implements CashbackCardService {
         String message;
 
         if(userId != null) {
-            Boolean result = cashbackCardRepository.deleteCashbackCardByUserId(userId);
-
-            message = messageSource.getMessage("delete.success", new String[]{}, locale);
+            if(cashbackCardRepository.existsByUserId(userId)) {
+                cashbackCardRepository.deleteCashbackCardByUserId(userId);
+                message = messageSource.getMessage("operation.success", new String[]{}, locale);
+                return ResponseDto.<Boolean>builder()
+                        .code(200)
+                        .responseData(true)
+                        .success(true)
+                        .message(message)
+                        .build();
+            }
+            message = messageSource.getMessage("not.found", new String[]{}, locale);
             return ResponseDto.<Boolean>builder()
-                    .code(200)
-                    .responseData(result)
-                    .success(true)
+                    .code(-2)
+                    .responseData(false)
+                    .success(false)
                     .message(message)
                     .build();
         }
